Svelte5 2 (#758)
* part-1

* comments-not-config

* helpermenu

* link-buttons

* modals

* tweaks

* alert

* whosGotWhat

* quickExtract

* GuildLog

* Modal

* VirtualList

* global-top

* components-feedback

* svelte-table

* remove superelite msg

* remove potreport msg

* deepsource-1

* deepsource-2

* oops

* deepsource-3

+++ src/modules/profile/components/Components.svelte
@@ -1,24 +1,33 @@
 <script>
-  import { createEventDispatcher } from 'svelte';
   import sendEvent from '../../analytics/sendEvent';
   import quickExtract from '../../chrome/pageSwitcher/loader/quickExtract';
-  import LinkButtonBracketed from '../../common/LinkButtonBracketed.svelte';
+  import LinkBtnBracketed from '../../common/LinkBtnBracketed.svelte';
   import Count from './Count.svelte';
 
-  let quickDelete;
-  let rollup;
+  const { dispatchDelete, dispatchDelType, dispatchQuickDel } = $props();
 
-  const dispatch = createEventDispatcher();
+  let quickDelete = $state();
+  let rollup = $state();
+
+  let quickDelHidden = true;
+  let countHidden = true;
+  let deleteHidden = true;
 
   function enableQuickDel() {
-    sendEvent('components', 'enableQuickDel');
-    quickDelete = true;
-    dispatch('enableQuickDel');
+    if (quickDelHidden) {
+      quickDelHidden = false;
+      sendEvent('components', 'enableQuickDel');
+      quickDelete = true;
+      dispatchQuickDel();
+    }
   }
 
   function countComponents() {
-    sendEvent('components', 'countComponents');
-    rollup = true;
+    if (countHidden) {
+      countHidden = false;
+      sendEvent('components', 'countComponents');
+      rollup = true;
+    }
   }
 
   function insertQuickExtract() {
@@ -27,41 +36,41 @@
   }
 
   function deleteAllVisible() {
-    sendEvent('components', 'deleteAllVisible');
-    dispatch('deleteAllVisible');
+    if (deleteHidden) {
+      deleteHidden = false;
+      sendEvent('components', 'deleteAllVisible');
+      dispatchDelete();
+    }
   }
 </script>
 
 <div class="fshCenter">
   {#if !quickDelete}
     <div>
-      <LinkButtonBracketed on:click|once={enableQuickDel}
-        >Enable Quick Del</LinkButtonBracketed
-      >
+      <LinkBtnBracketed onclick={enableQuickDel}>
+        Enable Quick Del
+      </LinkBtnBracketed>
     </div>
   {/if}
   {#if !rollup}
     <div>
-      <LinkButtonBracketed on:click|once={countComponents}
-        >Count Components</LinkButtonBracketed
-      >
+      <LinkBtnBracketed onclick={countComponents}>
+        Count Components
+      </LinkBtnBracketed>
     </div>
   {:else}
-    <Count on:delType />
+    <Count {dispatchDelType} />
   {/if}
   <div>
-    <LinkButtonBracketed on:click={insertQuickExtract}
-      >Quick Extract</LinkButtonBracketed
-    >
+    <LinkBtnBracketed onclick={insertQuickExtract}>
+      Quick Extract
+    </LinkBtnBracketed>
   </div>
   {#if quickDelete}
     <div>
-      <LinkButtonBracketed
-        --button-color="red"
-        on:click|once={deleteAllVisible}
-      >
+      <LinkBtnBracketed --button-color="red" onclick={deleteAllVisible}>
         Delete All Visible
-      </LinkButtonBracketed>
+      </LinkBtnBracketed>
     </div>
   {/if}
 </div>
